+++ src/Ray.hpp
@@ -25,7 +25,7 @@ public:
     const Vector3D &direction() const { return _direction; }
 
 
-    Point3D at(double t) const { return _origin + _direction * t; }
+    Point3D at(float t) const { return _origin + _direction * t; }
 
 private:
     Point3D _origin;
